changes for front end test and add FileService

+++ user/src/main/java/onyx/user/service/UserInfoServiceImpl.java
@@ -9,13 +9,15 @@ import onyx.user.domain.entity.UserEntity;
 import onyx.user.domain.valueobject.Profile;
 import onyx.user.domain.valueobject.ProfileImage;
 import onyx.user.repository.UserRepository;
-import onyx.user.web.UserInfoUpdateRequest;
+import onyx.user.web.UserProfileUpdateRequest;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.security.core.parameters.P;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.File;
 import java.io.IOException;
+import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -34,29 +36,43 @@ public class UserInfoServiceImpl implements UserInfoService {
 
     @Transactional
     @Override
-    public void updateUserInfo(Long userId, UserInfoUpdateRequest request) throws IOException {
+    public void updateUserInfo(Long userId, UserProfileUpdateRequest request) throws IOException {
 
         UserEntity user = getUserById(userId);
 
-        String tempName = UUID.randomUUID().toString();
+        if (request.getProfileImage() != null && !request.getProfileImage().isEmpty()) {
+            String tempName = UUID.randomUUID().toString();
 
-        File file = FileUtil.convertMultipartFileToFile(request.getProfileImage());
+            File file = FileUtil.convertMultipartFileToFile(request.getProfileImage());
 
-        try {
-            FileInfo fileInfo = fileStorageStrategy.saveFile(file, uploadDir, tempName);
+            try {
+                FileInfo fileInfo = fileStorageStrategy.saveFile(file, uploadDir, tempName);
 
-            Profile profile =
-                    Profile.create(ProfileImage.create(fileInfo), request.getBio());
+                Profile profile =
+                        Profile.create(ProfileImage.create(fileInfo), request.getBio());
 
-            fileRepository.save(fileInfo);
-            user.updateNickName(request.getNickName());
-            user.updateProfile(profile);
+                fileRepository.save(fileInfo);
+                user.updateNickName(request.getNickName());
+                user.updateProfile(profile);
 
-        } finally {
-            if (file.exists()) {
-                file.delete();
+            } finally {
+                if (file.exists()) {
+                    file.delete();
+                }
             }
+        }else {
+            user.updateNickName(request.getNickName());
+
+            Profile currentProfile = user.getProfile();
+            currentProfile.updateBio(request.getBio());
+            user.updateProfile(currentProfile);
         }
+
+    }
+
+    @Override
+    public Optional<UserEntity> getUserInfo(Long userId) {
+        return userRepository.findById(userId);
     }
 
     private UserEntity getUserById(Long userId) {
